Support event listeners.
This adds support for event listeners  Unfortunately, the DOM is a
little frustrating in this regard in that you cannot easily see what
event listeners are registered for a given Element.  Therefore, the
approach taken in this library is to record the current event listener
set for a given Element.  This allows it to be easily replaced.
However, it does add a constraint at this time that at most one listner
for any given event can be added to a given element.