Improve IO handling
This pushes through a fairly major change to the manner in which I/O is
handled.  Specifically, rather than rely on custom action processors (as
before), we have a generic and open-ended processing system which allows
arbitrary methods to be put through as events.  To smooth the process,
have added a bunch of standard methods initially.  This includes, for
example, the ability to query some property of a given DOM element based
on its ID.